adding important,bookmarks,notifications and settings views routes and also enabling routing within the p page. also updating app.vue with the various routes and creating home.vue in components
