+++ client/src/components/home.vue
@@ -1,0 +1,20 @@
+<template>
+  <div>this is the landing page</div>
+  <button></button>
+</template>
+
+<style lang="scss" scoped>
+div {
+  width: 100vw;
+  height: 100vh;
+  background: linear-gradient(300deg, rgb(5, 171, 159), teal);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  h1 {
+    text-transform: capitalize;
+    font: 600 50px "Poppins", sans-serif;
+  }
+}
+</style>
